+++ utils.go
@@ -70,6 +70,10 @@ func recordMetric(
 			attributes = append(attributes, cfg.InstrumentAttributesGetter(ctx, method, query, args)...)
 		}
 		if err != nil {
+			if cfg.InstrumentErrorAttributesGetter != nil {
+				attributes = append(attributes, cfg.InstrumentErrorAttributesGetter(err)...)
+			}
+
 			if cfg.DisableSkipErrMeasurement && err == driver.ErrSkip {
 				attributes = append(attributes, queryStatusKey.String("ok"))
 			} else {